requirements.txt
altair==5.4.0
attrs==24.2.0
bcrypt==3.1.7
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
flake8==7.1.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
narwhals==1.4.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pillow==10.4.0
pluggy==1.5.0
protobuf==5.27.3
psycopg2==2.9.9
pyarrow==17.0.0
pycodestyle==2.12.1
pycparser==2.22
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.3.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.32
streamlit==1.37.1
tenacity==8.5.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
Werkzeug==3.0.3
